+++ interface.c
@@ -3,7 +3,7 @@
 #include "interface.h"
 #include "game.h"
 
-extern int score, high_score;
+extern unsigned int score, high_score;
 
 void iniciaInterface(void) {
     initscr(); /*Inicia a "tela do jogo"*/
@@ -19,8 +19,8 @@ void encerraInterface(void) {
 }
 
 /*Função responsável por imprimir a interface de jogo*/
-void printInterface(int matrix[SIZE][SIZE]) {
-    int i, j;
+void printInterface(short int matrix[SIZE][SIZE]) {
+    short int i, j;
     refresh(); /*Reinicia a tela para expor novos valores*/
     clear(); /*Limpa a tela*/
